requirements.txt
Django==2.0.1
djangorestframework==3.7.7
numpy
psycopg2==2.7.3.2
pytz==2017.3
scipy
